lambdas/pipelines/title_embeddings/requirements.txt
boto3>=1.28.0
opensearch-py>=2.0.0
requests-aws4auth>=1.2.0
